Code cleanup, scope reduction
Python code formatter 'Black' has been used to format the entire project, and will be used to format all future commits. Several options have been removed from the config, including the options to have battle modes other than accept_challenge. This is to simplify the codebase and reduce the scope of the project.